Cleanup index.js by separating routes and store into new files
