+++ src/store.js
@@ -1,0 +1,19 @@
+import { createStore, applyMiddleware } from 'redux'
+import thunk from 'redux-thunk'
+import { composeWithDevTools } from 'redux-devtools-extension'
+
+import { selectUserAccountInfo } from './actions/user'
+import { getContractEvents } from './actions/events'
+import { fetchParameters } from './actions/parameters'
+
+import reducers from './reducers'
+
+export const store = createStore(
+  reducers,
+  composeWithDevTools(applyMiddleware(thunk))
+)
+
+store.dispatch(selectUserAccountInfo())
+store.dispatch(getContractEvents())
+store.dispatch(fetchParameters())
+store.dispatch(fetchParameters())
